Add mapreduce_single function
Since the demise of `r_promote` in #22825, there is now a type-instability in `mapreduce` if the operator does not give an element of the same type as the input. This arose during my implementation of Kahan summation using a reduction operator, see: https://github.com/JuliaMath/KahanSummation.jl/pull/7

This adds a `mapreduce_single` function which defines what the result should be in these cases.